Fix the weirdness that is hpack_setlimit.
To create a HPACK state object, simply allocate (and fill with zeroes) a hpack_t and pass it to hpack_init; the second argument is the initial upper limit on the dynamic table size (normally, this is obtained from an HTTP2 SETTINGS frame). If the state object is an encoder, the application may choose to set a lower limit on the dynamic table; this is done by calling hpack_setlimit with the new value. If an HTTP2 SETTINGS arrives with a new upper limit, the application should update the limit_upper field of the state object.

Simple.