Fix handling of continuation frames.
For some reason, the designers of HTTP 2 though it would be a good idea
to make the HEADERS frame carry the END_STREAM flag, even though it may
be followed by CONTINUATION frames. A header block fragment ends the
stream but does not end the block itself? Hell yea. Anyway, the obvious
solution is to ignore END_STREAM on HEADERS without END_HEADERS and
instead carry it through to the last CONTINUATION (the one with
END_HEADERS set.) Also, cno_frame_write should not erase END_STREAM
from HEADERS frames when splitting them.