[breaking] handle HTTP/1.x pipelining more consistently.
Previously, a persistent stream 1 was created on which multiple messages
could be received or sent. This was in contrast to HTTP 2, where each
stream would only carry one request-response pair.

The new behavior is to
  1. create stream 1 when a request is received or sent;
  2. destroy stream 1 when the transmission of the request is finished
     and a response is sent or received;
  3. *return CNO_ERRNO_WOULD_BLOCK from cno_connection_data_received
     and cno_write_message if an attempt is made to reuse stream 1
     but it is still active*.

The application is expected to handle CNO_ERRNO_WOULD_BLOCK by waiting
until (for a client) the response is completely consumed or (for a
server) a response is finalized. When that's done, (for a client)
call cno_write_message again, or (for a server) feed an empty buffer
to cno_connection_data_received (to handle the data that has been
passed to the previous call of this function.)

In theory, client applications should require no modifications as
cno_write_message may already return CNO_ERRNO_WOULD_BLOCK to indicate
that the limit on concurrent HTTP 2 streams has been reached. In both
cases the proper response is to wait for on_stream_end.

TODO support this in python bindings.